refactored summary data calc to server, added tabs for summary data, vis, and experimental details
