+++ src/lib/RawDataTable.svelte
@@ -33,7 +33,6 @@
     });
 
     $: dataset = filterDataset(selected)
-    $: console.log(dataset)
 
 </script>
